pyproject.toml
[project]
name = "parsehubbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiocache>=0.12.3",
    "apscheduler==3.10.4",
    "dynrender-skia>=0.2.5",
    "httpx>=0.28.1",
    "kurigram>=2.2.7",
    "loguru>=0.6.0,<0.7.0",
    "lxml-html-clean>=0.4.1",
    "markdown>=3.7",
    "parsehub>=1.2.37",
    "pillow>=11.3.0",
    "pillow-heif>=1.1.1",
    "python-dotenv>=1.0.1",
    "telegraph>=2.2.0",
    "tgcrypto>=1.2.5",
]

[tool.ruff]
line-length = 140

[tool.ruff.lint]
select = [
    "E", # pycodestyle 错误检查
    "W", # pycodestyle 警告检查
    "F", # pyflakes 错误检查
    "I", # isort 导入排序
    "B", # flake8-bugbear 常见错误检查
    "C4", # flake8-comprehensions 列表/字典推导式检查
    "UP", # pyupgrade 自动升级语法
]
ignore = [
    "B008", # 不在参数默认值中执行函数调用
    "C901", # 函数复杂度过高
]

